Move in-app entities to CommonIAP
Except those having app-specific behavior:

- AppFeatureRequiring extensions
- FallbackReceiptReader
- IAPManager
