+++ Passepartout/Library/Sources/CommonIAP/Domain/AppFeatureRequiring.swift
