+++ Passepartout/Library/Sources/CommonLibrary/Extensions/IAPManager+Verify.swift
