+++ Passepartout/Library/Sources/CommonLibrary/Shared.swift
@@ -23,6 +23,7 @@
 //  along with Passepartout.  If not, see <http://www.gnu.org/licenses/>.
 //
 
+@_exported import CommonIAP
 import Foundation
 import PassepartoutKit
